requirements.txt
pyfiglet
Pillow
psutil